\name{reader}
\alias{reader}
\docType{data}
\encoding{UTF-8}
\title{Who is a Regular Reader?}
\description{
The \code{reader} data contains information on the reading behaviour of women refering to a specific woman's journal.
}
\usage{data(reader)}
\format{
  A data frame with 48 observations on the following 5 variables. Considering the weighting variable "Freq", there are 941 observations. 
  \describe{
    \item{\code{RegularReader}}{Is the woman a regular reader? yes = 1, no = 0}
    \item{\code{Working}}{Is the woman working? yes = 1, no = 0}
    \item{\code{Age}}{Age of the woman in categories (18--29 years = 1, 30--39 = 2, 40--49 = 3)}
    \item{\code{Education}}{Level of education. L1 = 11, L2 = 12, L3 = 13, L4 = 14}
    \item{\code{Freq}}{Frequency of the observation}
  }
}
\source{Fahrmeir, Hamerle and Tutz (1996),
  Multivariate statistische Verfahren,
  Berlin: de Gruyter  
  }

\keyword{datasets}
